Refactor method `Table.set_edges` for readability
From all close points stored in `i`, `j` or `k` only the first one was used
therefore the new function `find_close_point` only returns the first match if
any.

Similar code is hoisted out of loops and branches and the `while` loops
are more clearly expressed with `for` loops.