Simplify and fix `Table.set_span`
In `Table.set_span` there are a bunch of redundant checks for borders and
number of bordered edges that this commit simplifies. Additionally the
vertical and horizontal spans are independent of each other, therefore these
checks can be untangled.

I think that this actually uncovered a bug in the old implementation.
For the cases where both horizontally and vertically only one of both edges
are bordered there are differences with this commit.

| left  | right | top   | bottom | vspan/hspan before | vspan/hspan after |
| ----- | ----- | ---   | ------ | ------------------ | ----------------- |
| true  | false | true  | false  | false              | true              |
| true  | false | false | true   | false              | true              |
| false | true  | true  | false  | false              | true              |
| false | true  | false | true   | false              | true              |

This change in behaviour of the `Table.set_span` method is a bug fix.
I guess in the original code it was missed that a cell could both
horizontally and vertically span to other cells if the two bordered
edges are not on the same direction.